# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Product Search RAG API",
    "version": "1.0.0",
    "description": "API for product recommendation and retrieval using RAG with SAP AI Core Document Grounding and GenAI Hub."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/retrieveData": {
      "post": {
        "summary": "Retrieve product recommendations or information",
        "description": "Accepts a user query and returns product recommendations or information based on retrieved context from SAP AI Core Document Grounding.",
        "operationId": "retrieveData",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "User's product-related question or search query."
                  }
                },
                "required": ["query"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response with product recommendations or information.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "description": "The answer generated by the RAG model, possibly structured or plain text.",
                      "oneOf": [
                        {"type": "string"},
                        {"type": "object"}
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or no result found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {"type": "string"}
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized access.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {"type": "string"}
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {"type": "string"}
                  }
                }
              }
            }
          }
        },
        "security": [
          {"bearerAuth": []}
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}